Разделил данные по модулям
