+++ tests/functional/testdata/test_param/genre_param.py
@@ -1,0 +1,12 @@
+from typing import List
+
+from .dto_param import AbstractDTO
+
+
+class Genre(AbstractDTO):
+    id: str
+    name: str | None
+
+
+class GenresList(AbstractDTO):
+    genres: List[Genre] | None
